+++ src/redis_hash.h
@@ -30,7 +30,7 @@ class Hash : public SubKeyScanner {
   rocksdb::Status SetNX(const Slice &user_key, const Slice &field, Slice value, int *ret);
   rocksdb::Status Delete(const Slice &user_key, const std::vector<Slice> &fields, int *ret);
   rocksdb::Status IncrBy(const Slice &user_key, const Slice &field, int64_t increment, int64_t *ret);
-  rocksdb::Status IncrByFloat(const Slice &user_key, const Slice &field, float increment, float *ret);
+  rocksdb::Status IncrByFloat(const Slice &user_key, const Slice &field, double increment, double *ret);
   rocksdb::Status MSet(const Slice &user_key, const std::vector<FieldValue> &field_values, bool nx, int *ret);
   rocksdb::Status MGet(const Slice &user_key,
                        const std::vector<Slice> &fields,
